--- solace-candidate-assignment-main/src/type/type.ts
@@ -1,10 +1,0 @@
-import { InferSelectModel } from 'drizzle-orm';
-import { advocates } from '@/db/schema';
-
-export type Advocate = InferSelectModel<typeof advocates>;
-
-export type HeaderProps = {
-  searchTerm: string;
-  onSearchChange: (value: string) => void;
-  onReset: () => void;
-};
